refactor: extract harbor prop placement to harbor_props.js and use new helpers in harbor.js
