+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/worldgen/town/harbor.js
@@ -1,8 +1,8 @@
 import { getGameData } from "../../utils/access.js";
 import { getTileDefByKey } from "../../data/tile_lookup.js";
 import { prepareHarborZone } from "./harbor_zone.js";
-import { propTypeFromCode, rectMostlyInHarborMask, rectOverlapsAny } from "./harbor_util.js";
 import { placeHarborBoatsAndEnsureAccess } from "./harbor_boats.js";
+import { placeDockPropsForHarbor, placeHarborWarehousesForHarbor } from "./harbor_props.js";
 
 // Re-export prepareHarborZone so callers can continue to import it from this module.
 export { prepareHarborZone };
@@ -466,141 +466,9 @@ export function placeHarborPrefabs(ctx, buildings, W, H, gate, plaza, rng, stamp
       placeHarborBoatsAndEnsureAccess(ctx, buildings, harborMask, harborDir, pierMask, W, H, WATER, gate, rng, PFB);
     }
 
-    // Simple dock props: reuse existing props (CRATE/BARREL/LAMP) along the harbor edge.
-      function placeDockProps() {
-        const bandCoords = [];
-        for (let y = 1; y < H - 1; y++) {
-          for (let x = 1; x < W - 1; x++) {
-            if (!harborMask[y][x]) continue;
-            if (ctx.map[y][x] !== ctx.TILES.FLOOR) continue;
-            // Avoid cluttering the gate bridge corridor with crates/barrels/lamps.
-            if (gateBridgeMask && gateBridgeMask[y] && gateBridgeMask[y][x]) continue;
-            bandCoords.push({ x, y });
-          }
-        }
-        if (!bandCoords.length) return;
-
-      // Rough heuristic: choose a handful of tiles along outermost band row/col to decorate.
-      const edgeCoords = [];
-      if (harborDir === "N") {
-        let minY = H;
-        for (const c of bandCoords) if (c.y < minY) minY = c.y;
-        for (const c of bandCoords) if (c.y === minY) edgeCoords.push(c);
-      } else if (harborDir === "S") {
-        let maxY = 0;
-        for (const c of bandCoords) if (c.y > maxY) maxY = c.y;
-        for (const c of bandCoords) if (c.y === maxY) edgeCoords.push(c);
-      } else if (harborDir === "W") {
-        let minX = W;
-        for (const c of bandCoords) if (c.x < minX) minX = c.x;
-        for (const c of bandCoords) if (c.x === minX) edgeCoords.push(c);
-      } else if (harborDir === "E") {
-        let maxX = 0;
-        for (const c of bandCoords) if (c.x > maxX) maxX = c.x;
-        for (const c of bandCoords) if (c.x === maxX) edgeCoords.push(c);
-      }
-
-      if (!edgeCoords.length) return;
-
-      // Thin sampling of edge tiles for simple dock accents
-      for (let i = 0; i < edgeCoords.length; i++) {
-        const c = edgeCoords[i];
-        const r = rng ? rng() : Math.random();
-        if (r < 0.15) {
-          _safeAddProp(ctx, W, H, c.x, c.y, "CRATE");
-        } else if (r < 0.30) {
-          _safeAddProp(ctx, W, H, c.x, c.y, "BARREL");
-        } else if (r < 0.36) {
-          // occasional lamp near docks
-          _safeAddProp(ctx, W, H, c.x, c.y, "LAMP");
-        }
-      }
-    }
-
-    // Harbor warehouses: pick small house prefabs and place them just inside the harbor band,
-    // away from the outermost row/column but still clearly \"at the harbor\".
-    function placeHarborWarehouses() {
-      if (!PFB || !Array.isArray(PFB.houses) || !PFB.houses.length) return;
-
-      // Filter for small-ish houses; we reuse them as generic warehouses visually.
-      const candidates = PFB.houses.filter(p => {
-        if (!p || !p.size) return false;
-        const w = p.size.w | 0;
-        const h = p.size.h | 0;
-        return w <= 9 && h <= 7;
-      });
-      if (!candidates.length) return;
-
-      // Build a simple list of candidate anchor positions inside the harbor band,
-      // but inset by 1 tile from the very edge so buildings don't overlap docks.
-      // Avoid using the gate bridge corridor as a warehouse anchor so the 4-wide
-      // approach from the gate remains visually open.
-      const bandCells = [];
-      for (let y = 1; y < H - 1; y++) {
-        for (let x = 1; x < W - 1; x++) {
-          if (!harborMask[y][x]) continue;
-          // Inset: require at least one tile from outer map border
-          if (x <= 1 || y <= 1 || x >= W - 2 || y >= H - 2) continue;
-          if (gateBridgeMask && gateBridgeMask[y] && gateBridgeMask[y][x]) continue;
-          bandCells.push({ x, y });
-        }
-      }
-      if (!bandCells.length) return;
-
-      // Try to place up to two warehouses.
-      const maxWarehouses = 2;
-      let placed = 0;
-      let attempts = 0;
-
-      while (placed < maxWarehouses && attempts++ < 40) {
-        const center = bandCells[Math.floor((rng ? rng() : Math.random()) * bandCells.length)];
-        if (!center) break;
-
-        const pref = candidates[Math.floor((rng ? rng() : Math.random()) * candidates.length)];
-        if (!pref || !pref.size) continue;
-        const bw = pref.size.w | 0;
-        const bh = pref.size.h | 0;
-
-        const bx = Math.max(1, Math.min(W - bw - 1, center.x - ((bw / 2) | 0)));
-        const by = Math.max(1, Math.min(H - bh - 1, center.y - ((bh / 2) | 0)));
-
-        // Ensure footprint stays mostly inside harbor band
-        if (!rectMostlyInHarborMask(harborMask, bx, by, bw, bh, W, H)) continue;
-        // Do not collide with existing buildings
-        if (rectOverlapsAny(buildings, bx, by, bw, bh)) continue;
-        // Ensure we are not stamping on water or other non-floor tiles
-        let anyBad = false;
-        for (let yy = by; yy < by + bh && !anyBad; yy++) {
-          for (let xx = bx; xx < bx + bw; xx++) {
-            if (yy <= 0 || yy >= H - 1 || xx <= 0 || xx >= W - 1) { anyBad = true; break; }
-            const t = ctx.map[yy][xx];
-            if (t !== ctx.TILES.FLOOR && t !== ctx.TILES.ROAD) { anyBad = true; break; }
-          }
-        }
-        if (anyBad) continue;
-
-        const ok = stampPrefab(ctx, pref, bx, by) || trySlipStamp(ctx, pref, bx, by, 2);
-        if (!ok) continue;
-
-        // Mark this building rect as harbor-tagged for AI and metadata
-        try {
-          const bRect = buildings.find(b => b && b.x === bx && b.y === by && b.w === bw && b.h === bh);
-          if (bRect) {
-            bRect.prefabCategory = bRect.prefabCategory || "harbor";
-            const existingTags = Array.isArray(bRect.prefabTags) ? bRect.prefabTags : [];
-            if (!existingTags.includes("harbor")) {
-              bRect.prefabTags = existingTags.concat(["harbor"]);
-            }
-          }
-        } catch (_) {}
-
-        placed++;
-      }
-    }
-
     carveHarborWaterAndPiers();
-    placeDockProps();
-    placeHarborWarehouses();
+    placeDockPropsForHarbor(ctx, W, H, harborMask, harborDir, gateBridgeMask, rng);
+    placeHarborWarehousesForHarbor(ctx, buildings, W, H, harborMask, gateBridgeMask, rng, PFB, stampPrefab, trySlipStamp);
 
     // Recompute harbor building list now that warehouses may have been stamped.
     try {
@@ -622,189 +490,5 @@ export function placeHarborPrefabs(ctx, buildings, W, H, gate, plaza, rng, stamp
     } catch (_) {}
 }
 
-function _safeAddProp(ctx, W, H, x, y, code) {
-  try {
-    if (!ctx || !ctx.townProps || !ctx.map) return false;
-    if (x <= 0 || y <= 0 || x >= W - 1 || y >= H - 1) return false;
-    const t = ctx.map[y][x];
-    // Allow dock props on generic outdoor ground, roads, and harbor piers/ship decks.
-    if (
-      t !== ctx.TILES.FLOOR &&
-      t !== ctx.TILES.ROAD &&
-      t !== ctx.TILES.PIER &&
-      t !== ctx.TILES.SHIP_DECK
-    ) {
-      return false;
-    }
-    if (Array.isArray(ctx.townProps) && ctx.townProps.some(p => p.x === x && p.y === y)) return false;
-    ctx.townProps.push({ x, y, type: propTypeFromCode(code), name: null });
-    return true;
-  } catch (_) {}
-  return false;
-}
 
-function _safeAddBoatProp(ctx, W, H, x, y, waterTile) {
-  try {
-    if (!ctx || !ctx.townProps || !ctx.map) return false;
-    if (x <= 0 || y <= 0 || x >= W - 1 || y >= H - 1) return false;
-    const t = ctx.map[y][x];
-    if (t !== waterTile) return false;
-    if (Array.isArray(ctx.townProps) && ctx.townProps.some(p => p.x === x && p.y === y)) return false;
-    ctx.townProps.push({ x, y, type: "boat", name: null });
-    return true;
-  } catch (_) {}
-  return false;
-}
-
-/**
- * Stamp a boat prefab on top of harbor water, converting selected water tiles
- * into bright ship deck and placing ship props (rails/masts/hatch).
- *
- * This is intentionally separate from the generic building prefab stamper:
- * - It allows stamping directly on water instead of FLOOR/ROAD.
- * - It uses SHIP_DECK tiles and ship_* props instead of WALL/FLOOR/etc.
- *
- * Preconditions:
- * - prefab.tiles is a size.w × size.h grid using codes:
- *   - "WATER"      -> leave underlying tile as-is (typically HARBOR_WATER)
- *   - "SHIP_DECK"  -> convert to deck tile
- *   - "SHIP_RAIL"  -> deck tile + ship_rail prop
- *   - "MAST"       -> deck tile + mast prop
- *   - "SHIP_HATCH" -> deck tile + ship_hatch prop
- */
-function _stampBoatPrefabOnWater(ctx, prefab, bx, by, W, H, harborMask, waterTile) {
-  if (!ctx || !prefab || !prefab.size || !Array.isArray(prefab.tiles)) return false;
-  const w = prefab.size.w | 0;
-  const h = prefab.size.h | 0;
-  if (!w || !h) return false;
-
-  const x0 = bx | 0;
-  const y0 = by | 0;
-  const x1 = x0 + w - 1;
-  const y1 = y0 + h - 1;
-  if (x0 <= 0 || y0 <= 0 || x1 >= W - 1 || y1 >= H - 1) return false;
-
-  // Validate row shapes
-  for (let yy = 0; yy < h; yy++) {
-    const row = prefab.tiles[yy];
-    if (!row || row.length !== w) return false;
-  }
-
-  // Resolve ship deck tile id (fallback to FLOOR if lookup fails).
-  let DECK = ctx.TILES.FLOOR;
-  try {
-    const td = getTileDefByKey("town", "SHIP_DECK") || null;
-    if (td && typeof td.id === "number") {
-      DECK = td.id | 0;
-    }
-  } catch (_) {}
-
-  // Ensure boat mask exists and matches map dims.
-  try {
-    const ok =
-      Array.isArray(ctx.townBoatMask) &&
-      ctx.townBoatMask.length === H &&
-      H > 0 &&
-      Array.isArray(ctx.townBoatMask[0]) &&
-      ctx.townBoatMask[0].length === W;
-    if (!ok) {
-      ctx.townBoatMask = Array.from({ length: H }, () => Array(W).fill(false));
-    }
-  } catch (_) {
-    ctx.townBoatMask = Array.from({ length: H }, () => Array(W).fill(false));
-  }
-  const boatMask = ctx.townBoatMask;
-
-  // Ensure props container.
-  try {
-    if (!Array.isArray(ctx.townProps)) ctx.townProps = [];
-  } catch (_) {}
-
-  // First pass: validate that all non-WATER codes sit on harbor water inside the mask.
-  for (let yy = 0; yy < h; yy++) {
-    const row = prefab.tiles[yy];
-    for (let xx = 0; xx < w; xx++) {
-      const codeRaw = row[xx];
-      const code = codeRaw ? String(codeRaw).toUpperCase() : "";
-      if (!code || code === "WATER") continue;
-      const tx = x0 + xx;
-      const ty = y0 + yy;
-      if (tx <= 0 || ty <= 0 || tx >= W - 1 || ty >= H - 1) return false;
-      if (harborMask && (!harborMask[ty] || !harborMask[ty][tx])) return false;
-      const t = ctx.map[ty][tx];
-      if (t !== waterTile) return false;
-    }
-  }
-
-  // Second pass: apply deck tiles and props.
-  for (let yy = 0; yy < h; yy++) {
-    const row = prefab.tiles[yy];
-    for (let xx = 0; xx < w; xx++) {
-      const codeRaw = row[xx];
-      const code = codeRaw ? String(codeRaw).toUpperCase() : "";
-      const tx = x0 + xx;
-      const ty = y0 + yy;
-
-      if (!code || code === "WATER") {
-        continue; // leave underlying water
-      }
-
-      // Decide which underlying tile to use for this ship cell.
-      // Default to deck; rails sit on edge tiles so the hull has a darker belt.
-      let tileId = DECK;
-      try {
-        if (code === "SHIP_EDGE") {
-          tileId = (ctx.TILES && ctx.TILES.SHIP_EDGE) || DECK;
-        } else if (code === "SHIP_RAIL") {
-          tileId = (ctx.TILES && ctx.TILES.SHIP_EDGE) || DECK;
-        }
-      } catch (_) {}
-
-      ctx.map[ty][tx] = tileId;
-      boatMask[ty][tx] = true;
-
-      // Only one prop per cell; skip if something already exists here.
-      let hasProp = false;
-      if (Array.isArray(ctx.townProps)) {
-        for (let i = 0; i < ctx.townProps.length; i++) {
-          const p = ctx.townProps[i];
-          if (p && p.x === tx && p.y === ty) {
-            hasProp = true;
-            break;
-          }
-        }
-      }
-      if (hasProp) continue;
-
-      if (code === "SHIP_RAIL") {
-        ctx.townProps.push({ x: tx, y: ty, type: "ship_rail", name: null });
-      } else if (code === "MAST") {
-        ctx.townProps.push({ x: tx, y: ty, type: "mast", name: null });
-      } else if (code === "SHIP_HATCH") {
-        ctx.townProps.push({ x: tx, y: ty, type: "ship_hatch", name: null });
-      } else if (code === "CRATE" || code === "BARREL" || code === "LAMP") {
-        // Generic cargo/light props on deck/edge tiles.
-        ctx.townProps.push({ x: tx, y: ty, type: propTypeFromCode(code), name: null });
-      }
-    }
-  }
-
-  // Record boat metadata for potential AI/interior use.
-  try {
-    const id = prefab && prefab.id ? String(prefab.id) : null;
-    if (id) {
-      if (!Array.isArray(ctx.townBoats)) ctx.townBoats = [];
-      ctx.townBoats.push({
-        id,
-        x: x0,
-        y: y0,
-        w,
-        h,
-        orientation: prefab.orientation || null
-      });
-    }
-  } catch (_) {}
-
-  return true;
-}
